Migrate ESLint to flat config and upgrade dependencies to latest LTS-safe versions
Finishes the in-progress ESLint flat-config migration (.eslintrc.js/.eslintignore
-> eslint.config.js) and bumps every dependency to the latest version within its
current major (Next 16.2.12, React 19.2.8, Tailwind 4.3.3, Jest 30.4.2, the
@typescript-eslint/OpenTelemetry families, etc.), skipping riskier major jumps
(TypeScript 7, ESLint 10, Babel 8, @vercel/analytics and speed-insights 2.x,
eslint-plugin-tailwindcss 4) for a dedicated follow-up.

Combining the ESLint bump with eslint-config-next's newer "next/typescript"
preset surfaced a duplicate @typescript-eslint plugin registration that newer
ESLint now rejects — fixed by registering the plugin exactly once in
eslint.config.js instead of letting both eslint-config-next and
typescript-eslint each register their own instance.

Also fixes the stale baseline-browser-mapping/caniuse-lite warning via
update-browserslist-db, and cleans up two next.config.mjs options the Next.js
bump made obsolete (instrumentationHook is default-on now; browserDebugInfoInTerminal
moved to logging.browserToTerminal).

Verified via yarn lint / tsc --noEmit / test / build, all clean.
