--- .eslintignore
@@ -1,2 +1,0 @@
-.next
-node_modules
