+++ .gitignore
@@ -11,6 +11,9 @@ coverage
 out/
 build
 
+# typescript
+*.tsbuildinfo
+
 # misc
 .DS_Store
 *.pem
