+++ .vscode/settings.json
@@ -14,13 +14,19 @@
     "statusBar.background": "#571d8f",
     "statusBar.foreground": "#e7e7e7",
     "statusBarItem.hoverBackground": "#7126b9",
-    "statusBarItem.remoteBackground": "#571d8f",
+    "statusBarItem.remoteBackground": "#9f6120",
     "statusBarItem.remoteForeground": "#e7e7e7",
-    "tab.activeBorder": "#7126b9",
     "titleBar.activeBackground": "#571d8f",
     "titleBar.activeForeground": "#e7e7e7",
     "titleBar.inactiveBackground": "#571d8f99",
-    "titleBar.inactiveForeground": "#e7e7e799"
+    "titleBar.inactiveForeground": "#e7e7e799",
+    "activityBarTop.activeBackground": "#7126b9",
+    "activityBarTop.background": "#7126b9",
+    "activityBarTop.foreground": "#e7e7e7",
+    "activityBarTop.inactiveForeground": "#e7e7e799",
+    "commandCenter.foreground": "#e7e7e7",
+    "statusBar.debuggingBackground": "#571d8f",
+    "statusBar.debuggingForeground": "#e7e7e7"
   },
   "peacock.color": "#571d8f",
   "i18n-ally.localesPaths": ["i18n/messages"],
