+++ jest.config.js
@@ -1,4 +1,4 @@
-import nextJest from "next/jest"
+import nextJest from "next/jest.js"
 
 const createJestConfig = nextJest({
   dir: "./",
@@ -9,6 +9,9 @@ const customJestConfig = {
   testEnvironment: "jest-environment-jsdom",
   modulePathIgnorePatterns: ["<rootDir>/dist/"],
   testPathIgnorePatterns: ["<rootDir>/e2e"],
+  // tsconfig.json sets baseUrl "." for root-relative imports (e.g. "lib/foo");
+  // mirror that here so Jest's resolver matches TypeScript's.
+  modulePaths: ["<rootDir>"],
 }
 
 export default createJestConfig(customJestConfig)
