+++ package.json
@@ -22,88 +22,88 @@
   "dependencies": {
     "@anthropic-ai/sdk": "^0.115.0",
     "@neondatabase/serverless": "^1.1.0",
-    "@next/bundle-analyzer": "^16.0.8",
+    "@next/bundle-analyzer": "^16.2.12",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/commit-analyzer": "13.0.1",
     "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^12.0.2",
-    "@semantic-release/npm": "^13.1.2",
-    "@semantic-release/release-notes-generator": "^14.1.0",
-    "@t3-oss/env-nextjs": "^0.13.8",
-    "@tailwindcss/postcss": "^4.1.17",
+    "@semantic-release/github": "^12.0.9",
+    "@semantic-release/npm": "^13.1.5",
+    "@semantic-release/release-notes-generator": "^14.1.1",
+    "@t3-oss/env-nextjs": "^0.13.11",
+    "@tailwindcss/postcss": "^4.3.3",
     "@u-wave/react-vimeo": "^0.9.12",
     "@vercel/analytics": "^1.6.1",
     "@vercel/otel": "^2.1.0",
     "@vercel/speed-insights": "^1.3.1",
     "class-variance-authority": "^0.7.1",
-    "flowbite": "^4.0.1",
-    "flowbite-react": "^0.12.12",
+    "flowbite": "^4.0.2",
+    "flowbite-react": "^0.12.17",
     "motion": "^12.43.0",
-    "next": "^16.0.8",
+    "next": "^16.2.12",
     "next-compose-plugins": "^2.2.1",
-    "next-intl": "^4.5.8",
+    "next-intl": "^4.13.4",
     "next-themes": "^0.4.4",
-    "react": "^19.2.1",
-    "react-dom": "^19.2.1",
-    "react-icons": "^5.4.0",
-    "tailwind-merge": "^3.4.0",
-    "zod": "^4.1.13"
+    "react": "^19.2.8",
+    "react-dom": "^19.2.8",
+    "react-icons": "^5.7.0",
+    "tailwind-merge": "^3.6.0",
+    "zod": "^4.4.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.28.5",
-    "@babel/plugin-syntax-flow": "^7.26.0",
-    "@babel/plugin-transform-optional-chaining": "^7.28.5",
-    "@babel/plugin-transform-react-jsx": "^7.25.9",
-    "@eslint/eslintrc": "^3.3.3",
-    "@jest/globals": "^30.2.0",
-    "@opentelemetry/api": "1.9.0",
-    "@opentelemetry/api-logs": "^0.208.0",
-    "@opentelemetry/instrumentation": "^0.208.0",
-    "@opentelemetry/resources": "2.2.0",
-    "@opentelemetry/sdk-logs": "^0.208.0",
-    "@opentelemetry/sdk-metrics": "^2.2.0",
-    "@opentelemetry/sdk-node": "0.208.0",
-    "@opentelemetry/sdk-trace-base": "^2.2.0",
-    "@opentelemetry/sdk-trace-node": "2.2.0",
-    "@opentelemetry/semantic-conventions": "1.38.0",
+    "@babel/core": "^7.29.7",
+    "@babel/plugin-syntax-flow": "^7.29.7",
+    "@babel/plugin-transform-optional-chaining": "^7.29.7",
+    "@babel/plugin-transform-react-jsx": "^7.29.7",
+    "@eslint/eslintrc": "^3.3.6",
+    "@jest/globals": "^30.4.1",
+    "@opentelemetry/api": "1.9.1",
+    "@opentelemetry/api-logs": "^0.221.0",
+    "@opentelemetry/instrumentation": "^0.221.0",
+    "@opentelemetry/resources": "2.10.0",
+    "@opentelemetry/sdk-logs": "^0.221.0",
+    "@opentelemetry/sdk-metrics": "^2.10.0",
+    "@opentelemetry/sdk-node": "0.221.0",
+    "@opentelemetry/sdk-trace-base": "^2.10.0",
+    "@opentelemetry/sdk-trace-node": "2.10.0",
+    "@opentelemetry/semantic-conventions": "1.43.0",
     "@testing-library/dom": "^10.0.0",
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.1.0",
+    "@testing-library/jest-dom": "6.9.1",
+    "@testing-library/react": "^16.3.2",
     "@total-typescript/ts-reset": "^0.6.1",
     "@types/jest": "^30.0.0",
-    "@types/node": "^24.10.2",
-    "@types/react": "^19.2.7",
+    "@types/node": "^24.13.3",
+    "@types/react": "^19.2.17",
     "@types/react-dom": "^19.2.3",
-    "@typescript-eslint/eslint-plugin": "^8.49.0",
-    "@typescript-eslint/parser": "^8.49.0",
+    "@typescript-eslint/eslint-plugin": "^8.65.0",
+    "@typescript-eslint/parser": "^8.65.0",
     "all-contributors-cli": "^6.26.1",
-    "autoprefixer": "^10.4.22",
+    "autoprefixer": "^10.5.4",
     "cross-env": "^10.1.0",
-    "eslint": "^9.39.1",
-    "eslint-config-next": "16.0.8",
+    "eslint": "^9.39.5",
+    "eslint-config-next": "16.2.12",
     "eslint-config-prettier": "^10.1.8",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-import-resolver-typescript": "^4.4.5",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-react": "7.37.5",
-    "eslint-plugin-tailwindcss": "^3.17.5",
+    "eslint-plugin-tailwindcss": "^3.18.3",
     "fetch-mock": "^12.6.0",
-    "jest": "^30.2.0",
-    "jest-environment-jsdom": "^30.2.0",
+    "jest": "^30.4.2",
+    "jest-environment-jsdom": "^30.4.1",
     "mkdirp": "^3.0.1",
     "npm-only-allow": "^1.2.6",
     "patch-package": "^8.0.1",
-    "postcss": "^8.5.6",
+    "postcss": "^8.5.24",
     "postinstall-postinstall": "^2.1.0",
-    "prettier": "3.7.4",
-    "prettier-plugin-tailwindcss": "^0.7.2",
-    "semantic-release": "^25.0.2",
-    "tailwindcss": "^4.1.17",
-    "ts-jest": "^29.4.6",
+    "prettier": "3.9.6",
+    "prettier-plugin-tailwindcss": "^0.7.4",
+    "semantic-release": "^25.0.8",
+    "tailwindcss": "^4.3.3",
+    "ts-jest": "^29.4.12",
     "tsc": "^2.0.4",
     "typescript": "5.9.3",
-    "typescript-eslint": "8.49.0",
-    "webpack": "5.103.0"
+    "typescript-eslint": "8.65.0",
+    "webpack": "5.109.2"
   },
   "engines": {
     "node": ">=22.0.0"
